Major Rework of the database system
I completely reworked the class, instead of being a copy of a HashMap using a HashMap object, it now uses a file system as a database instead. I made it so it has a designated file to read and write to. Now it works by first loading all previous data from the file into a HashMap that the user can interact with through methods. The user can still add, update, remove, or read information from the Database through the HashMap. Every time the HashMap is updated, the file gets updated as well to store the data for later on.